feat: 성능 최적화 및 초기 로딩 시간 단축
